Extract location detection into own module
This will make it easier to write unit tests.
