+++ src/app/api/v1/location-autocomplete/route.ts
@@ -3,13 +3,13 @@
  * file, You can obtain one at http://mozilla.org/MPL/2.0/. */
 
 import { NextResponse, NextRequest } from "next/server";
-import uFuzzy from "@leeoniya/ufuzzy";
 
 import { RelevantLocation } from "./types";
 // The location autocomplete data will be created during the build step.
 // @ts-ignore-next-line
 // eslint-disable-next-line import/no-unresolved
 import locationData from "../../../../../locationAutocompleteData.json";
+import { getRelevantLocations } from "./getRelevantLocations";
 
 export type MatchingLocations = Array<RelevantLocation> | [];
 
@@ -26,68 +26,6 @@ export interface SearchLocationResults {
   results: MatchingLocations;
 }
 
-function getLocationsByQuery(searchQuery: string) {
-  if (!locationData) {
-    throw new Error(
-      "No location data available: You may need to run `npm run create-location-data`.",
-    );
-  }
-
-  const locationNames = locationData.data.map((location: RelevantLocation) => {
-    const { n, s, a } = location;
-    const alternateNamesJoined = a ? a.join(" ") : "";
-    const countryCode = "USA";
-
-    return `${n} ${s} ${countryCode} ${alternateNamesJoined}`;
-  });
-
-  // For search options see: https://github.com/leeoniya/uFuzzy#options
-  const fuzzySearch = new uFuzzy({
-    intraMode: 1,
-    intraIns: 1,
-    interIns: 3,
-    intraSub: 1,
-    intraTrn: 1,
-    intraDel: 1,
-    // matches lowercase letters, whitespace or apostrophe
-    intraChars: "[a-z\\s\\d']",
-    interLft: 2,
-    interRgt: 0,
-  });
-
-  const locationIndexes = fuzzySearch.filter(locationNames, searchQuery);
-  if (!locationIndexes) {
-    return [];
-  }
-
-  const info = fuzzySearch.info(locationIndexes, locationNames, searchQuery);
-  const order = fuzzySearch.sort(info, locationNames, searchQuery);
-  const results = locationIndexes.map(
-    (locationIndex: number) =>
-      locationData.data[locationIndex] as RelevantLocation,
-  );
-
-  const resultsOrdered = order.map((orderIndex: number) => results[orderIndex]);
-  // Split and only sort the first x percentage of results by population.
-  // The motivation behind this is to improve the score for matches but not all
-  // of them in order to not move up weak ones.
-  const maxSortByPopulationThreshold = 0.75;
-  const locationSplitIndex = Math.ceil(
-    results.length * maxSortByPopulationThreshold,
-  );
-  const resultsSortedByPopulation = [
-    ...resultsOrdered
-      .slice(0, locationSplitIndex)
-      .sort(
-        (a: RelevantLocation, b: RelevantLocation) =>
-          Number(b.p ?? 0) - Number(a.p ?? 0),
-      ),
-    ...resultsOrdered.slice(locationSplitIndex + 1),
-  ];
-
-  return resultsSortedByPopulation;
-}
-
 function getLocationsResults({
   searchQuery,
   config = {
@@ -95,11 +33,17 @@ function getLocationsResults({
     maxResults: 5,
   },
 }: SearchLocationParams): SearchLocationResults {
+  if (!locationData) {
+    throw new Error(
+      "No location data available: You may need to run `npm run create-location-data`.",
+    );
+  }
+
   const { minQueryLength, maxResults } = config;
 
   const matchingLocations =
     searchQuery && searchQuery.length >= minQueryLength
-      ? getLocationsByQuery(searchQuery)
+      ? getRelevantLocations(searchQuery, locationData.data)
       : [];
 
   const locationsResults =
